@@ -1,47 +1,35 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using System.Reflection;
-using poi.Data;
+using Microsoft.Extensions.Hosting;
 using Newtonsoft.Json;
-using Microsoft.AspNetCore.Rewrite;
+using poi.Data;
 
 namespace IntegrationTests
 {
     public class Startup
     {
-        public Startup(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
+        public Startup(IConfiguration configuration) 
+            => Configuration = configuration;
 
         public IConfiguration Configuration { get; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc()
-                .AddJsonOptions(options =>
+            services.AddControllers()
+                .AddNewtonsoftJson((options =>
                 {
                     options.SerializerSettings.Formatting = Formatting.Indented;
-                });
+                }));
 
-                // Create a new service provider.
-                var serviceProvider = new ServiceCollection()
-                    .AddEntityFrameworkInMemoryDatabase()
-                    .BuildServiceProvider();
-
-                // Add a database context (ApplicationDbContext) using an in-memory
-                // database for testing.
-                services.AddDbContext<POIContext>(options =>
+            // Add a database context (ApplicationDbContext) using an in-memory
+            // database for testing.
+            services
+                .AddEntityFrameworkInMemoryDatabase()
+                .AddDbContext<POIContext>((serviceProvider, options) =>
                 {
                     options.UseInMemoryDatabase("InMemoryDbForTesting");
                     options.UseInternalServiceProvider(serviceProvider);
@@ -49,15 +37,10 @@ namespace IntegrationTests
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
-            {
                 app.UseDeveloperExceptionPage();
-                app.UseBrowserLink();
-            }
-
-            app.UseMvc();
         }
     }
 }
